Add usePositioning.js as custom hook for geolocation
